Modify all functions when considering q == NULL
Consider the situation if q == NULL for all functions.

If queue doesn't contain any elements, or malloc() returns NULL,
the rest of the instructions can be error. Pointers are not able
to access variables or become dangling pointers.

As a result, add conditions [ if (q == NULL) ] in every functions
and [ return false; ]. Make sure that there are no dangling pointers
exists.